+++ CHANGELOG.md
@@ -17,6 +17,7 @@
  * [#50](https://github.com/iliekturtles/uom/issues/50) Implement `Eq`.
  * [#18](https://github.com/iliekturtles/uom/issues/18) Implement `PartialOrd`.
  * [#51](https://github.com/iliekturtles/uom/issues/51) Implement `Ord`.
+ * [#15](https://github.com/iliekturtles/uom/issues/15) Implement `Product`.
  * [#26](https://github.com/iliekturtles/uom/issues/26) Implement `num::Zero`.
  * [#35](https://github.com/iliekturtles/uom/issues/35) Implement `num::Saturating`.
  * [#37](https://github.com/iliekturtles/uom/issues/35) Implement `serde::Serialize` and